Refactor more tool test verification code into galaxy-lib.
Should enable greater reuse from outside the test framework (e.g. standalone scripts and Planemo).
